# tsdb: sample chunk info from tsdb index to limit the amount of chunkrefs we read from index