--- a/pkg/storage/stores/tsdb/querier_test.go
+++ b/pkg/storage/stores/tsdb/querier_test.go
@@ -2,6 +2,7 @@ package tsdb
 
 import (
 	"context"
+	"math"
 	"testing"
 	"time"
 
@@ -113,12 +114,12 @@ func TestQueryIndex(t *testing.T) {
 	)
 
 	require.True(t, p.Next())
-	_, err = reader.Series(p.At(), &ls, &chks)
+	_, err = reader.Series(p.At(), 0, math.MaxInt64, &ls, &chks)
 	require.Nil(t, err)
 	require.Equal(t, cases[0].labels.String(), ls.String())
 	require.Equal(t, cases[0].chunks, chks)
 	require.True(t, p.Next())
-	_, err = reader.Series(p.At(), &ls, &chks)
+	_, err = reader.Series(p.At(), 0, math.MaxInt64, &ls, &chks)
 	require.Nil(t, err)
 	require.Equal(t, cases[1].labels.String(), ls.String())
 	require.Equal(t, cases[1].chunks, chks)