--- a/pkg/storage/stores/tsdb/single_file_index.go
+++ b/pkg/storage/stores/tsdb/single_file_index.go
@@ -102,12 +102,7 @@ func (i *TSDBIndex) SetChunkFilterer(chunkFilter chunk.RequestChunkFilterer) {
 
 // fn must NOT capture it's arguments. They're reused across series iterations and returned to
 // a pool after completion.
-func (i *TSDBIndex) forSeries(
-	ctx context.Context,
-	shard *index.ShardAnnotation,
-	fn func(labels.Labels, model.Fingerprint, []index.ChunkMeta),
-	matchers ...*labels.Matcher,
-) error {
+func (i *TSDBIndex) forSeries(ctx context.Context, shard *index.ShardAnnotation, from model.Time, through model.Time, fn func(labels.Labels, model.Fingerprint, []index.ChunkMeta), matchers ...*labels.Matcher) error {
 	p, err := PostingsForMatchers(i.reader, shard, matchers...)
 	if err != nil {
 		return err
@@ -123,7 +118,7 @@ func (i *TSDBIndex) forSeries(
 	}
 
 	for p.Next() {
-		hash, err := i.reader.Series(p.At(), &ls, &chks)
+		hash, err := i.reader.Series(p.At(), int64(from), int64(through), &ls, &chks)
 		if err != nil {
 			return err
 		}
@@ -143,32 +138,23 @@ func (i *TSDBIndex) forSeries(
 }
 
 func (i *TSDBIndex) GetChunkRefs(ctx context.Context, userID string, from, through model.Time, res []ChunkRef, shard *index.ShardAnnotation, matchers ...*labels.Matcher) ([]ChunkRef, error) {
-	queryBounds := newBounds(from, through)
 	if res == nil {
 		res = ChunkRefsPool.Get()
 	}
 	res = res[:0]
 
-	if err := i.forSeries(ctx, shard,
-		func(ls labels.Labels, fp model.Fingerprint, chks []index.ChunkMeta) {
-			// TODO(owen-d): use logarithmic approach
-			for _, chk := range chks {
+	if err := i.forSeries(ctx, shard, from, through, func(ls labels.Labels, fp model.Fingerprint, chks []index.ChunkMeta) {
+		for _, chk := range chks {
 
-				// current chunk is outside the range of this request
-				if !Overlap(queryBounds, chk) {
-					continue
-				}
-
-				res = append(res, ChunkRef{
-					User:        userID, // assumed to be the same, will be enforced by caller.
-					Fingerprint: fp,
-					Start:       chk.From(),
-					End:         chk.Through(),
-					Checksum:    chk.Checksum,
-				})
-			}
-		},
-		matchers...); err != nil {
+			res = append(res, ChunkRef{
+				User:        userID, // assumed to be the same, will be enforced by caller.
+				Fingerprint: fp,
+				Start:       chk.From(),
+				End:         chk.Through(),
+				Checksum:    chk.Checksum,
+			})
+		}
+	}, matchers...); err != nil {
 		return nil, err
 	}
 
@@ -176,27 +162,20 @@ func (i *TSDBIndex) GetChunkRefs(ctx context.Context, userID string, from, throu
 }
 
 func (i *TSDBIndex) Series(ctx context.Context, _ string, from, through model.Time, res []Series, shard *index.ShardAnnotation, matchers ...*labels.Matcher) ([]Series, error) {
-	queryBounds := newBounds(from, through)
 	if res == nil {
 		res = SeriesPool.Get()
 	}
 	res = res[:0]
 
-	if err := i.forSeries(ctx, shard,
-		func(ls labels.Labels, fp model.Fingerprint, chks []index.ChunkMeta) {
-			// TODO(owen-d): use logarithmic approach
-			for _, chk := range chks {
-				if Overlap(queryBounds, chk) {
-					// this series has at least one chunk in the desired range
-					res = append(res, Series{
-						Labels:      ls.Copy(),
-						Fingerprint: fp,
-					})
-					break
-				}
-			}
-		},
-		matchers...); err != nil {
+	if err := i.forSeries(ctx, shard, from, through, func(ls labels.Labels, fp model.Fingerprint, chks []index.ChunkMeta) {
+		if len(chks) == 0 {
+			return
+		}
+		res = append(res, Series{
+			Labels:      ls.Copy(),
+			Fingerprint: fp,
+		})
+	}, matchers...); err != nil {
 		return nil, err
 	}
 
@@ -233,21 +212,18 @@ func (i *TSDBIndex) Identifier(string) SingleTenantTSDBIdentifier {
 }
 
 func (i *TSDBIndex) Stats(ctx context.Context, userID string, from, through model.Time, acc IndexStatsAccumulator, shard *index.ShardAnnotation, shouldIncludeChunk shouldIncludeChunk, matchers ...*labels.Matcher) error {
-	if err := i.forSeries(ctx, shard,
-		func(ls labels.Labels, fp model.Fingerprint, chks []index.ChunkMeta) {
-			// TODO(owen-d): use logarithmic approach
-			var addedStream bool
-			for _, chk := range chks {
-				if shouldIncludeChunk(chk) {
-					if !addedStream {
-						acc.AddStream(fp)
-						addedStream = true
-					}
-					acc.AddChunk(fp, chk)
+	if err := i.forSeries(ctx, shard, from, through, func(ls labels.Labels, fp model.Fingerprint, chks []index.ChunkMeta) {
+		var addedStream bool
+		for _, chk := range chks {
+			if shouldIncludeChunk(chk) {
+				if !addedStream {
+					acc.AddStream(fp)
+					addedStream = true
 				}
+				acc.AddChunk(fp, chk)
 			}
-		},
-		matchers...); err != nil {
+		}
+	}, matchers...); err != nil {
 		return err
 	}