@honganan 2023-03-13
Great! I'm looking forward to similar index query improvement for a long time.